big refactor; create models package, move com.accountmanager to com.models.account, move com.market to com.models.market
