+++ src/main/java/com/etl/finnhub/QuoteClient.java
@@ -2,9 +2,8 @@ package com.etl.finnhub;
 
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import com.market.Database;
-import com.market.TradeListener;
-import io.github.cdimascio.dotenv.Dotenv;
+import com.models.Database;
+import com.models.market.TradeListener;
 
 import java.io.IOException;
 import java.net.URI;
