+++ src/main/java/com/gui/AccountCard.java
@@ -1,7 +1,7 @@
 // file: com/gui/AccountCard.java
 package com.gui;
 
-import com.accountmanager.Account;
+import com.models.profile.Account;
 
 import javax.swing.*;
 import java.awt.*;
