+++ src/main/java/com/gui/AccountPanel.java
@@ -1,6 +1,6 @@
 package com.gui;
 
-import com.accountmanager.Account;
+import com.models.profile.Account;
 
 import javax.swing.*;
 import java.awt.*;
