+++ src/main/java/com/models/ModelFacade.java
@@ -1,0 +1,7 @@
+package com.models;
+
+import com.models.market.Order;
+
+public class ModelFacade {
+    Order order;
+}
