+++ src/main/java/com/models/market/Stock.java
@@ -1,4 +1,4 @@
-package com.market;
+package com.models.market;
 
 public class Stock extends TradeItem {
