+++ src/main/java/com/models/market/TradeListener.java
@@ -1,4 +1,4 @@
-package com.market;
+package com.models.market;
 
 public interface TradeListener {
    public void onTrade(String symbol, double price);
