+++ src/main/java/com/models/market/TradeManager.java
@@ -1,4 +1,4 @@
-package com.market;
+package com.models.market;
 
 public class TradeManager {
 }
