+++ src/main/java/com/models/profile/Watchlist.java
@@ -1,6 +1,6 @@
-package com.accountmanager;
+package com.models.profile;
 
-import com.market.*;
+import com.models.market.TradeItem;
 
 
 import java.util.ArrayList;
