+++ src/main/java/com/tools/BuildSampleDb.java
@@ -1,6 +1,6 @@
 package com.tools;
 
-import com.market.Database;
+import com.models.Database;
 import java.time.*;
 import java.util.*;
